교보 크롤러, 전처리
